+++ controllers/categoria.js
@@ -1,4 +1,7 @@
 import categoria from '../models/categoria.js';
+import util from './util.js';
+
+const tabela = 'categoria';
 
 const listar = async (req, res) => {
     const dados = await categoria.findAll();
@@ -16,22 +19,62 @@ const selecionar = async (req, res) => {
 }
 
 const inserir = async (req, res) => {
+    const requiredColumns = await util.requiredColumns(tabela);
+    const permittedColumns = await util.permittedColumns(tabela);
+    const data = util.filterObjectKeys(req.body, permittedColumns);
+    if (util.keysMatch(data, requiredColumns) === false) {
+        return res.status(400).json({
+            mensagem: "Dados obrigatórios não informados",
+            obrigatorios: requiredColumns,
+            informados: Object.keys(data)
+        });
+    }
+    console.log(data);
     await categoria.create(req.body)
         .then(result => res.status(201).json(result))
         .catch(err => res.status(400).json(err));
 }
 
 const alterar = async (req, res) => {
+    const permittedColumns = await util.permittedColumns(tabela);
+    const data = util.filterObjectKeys(req.body, [...permittedColumns, 'id']);
+    if (Object.keys(data).length == 0) {
+        return res.status(400).json({
+            mensagem: "Nenhum dado informado para atualização",
+            permitidos: permittedColumns,
+            informados: Object.keys(data)
+        });
+    }
+    // Verifica se o ID é um número
+    if (!util.isNumber(req.params.id)) {
+        return res.status(400).json({ message: "ID inválido" });
+    }
+    // Verifica se o ID existe
+    const categoriaExistente = await categoria.findByPk(req.params.id);
+    if (!categoriaExistente) {
+        return res.status(404).json({ message: "Categoria não encontrada" });
+    }
+    // Atualiza os dados
     await categoria.update(req.body, {
         where: {
             id: req.params.id
         }
     })
         .then(result => res.status(200).json(result))
-        .catch(err => res.status(400).json(err));
+        .catch(err => res.status(500).json(err));
 }
 
 const excluir = async (req, res) => {
+    // Verifica se o ID é um número
+    if (!util.isNumber(req.params.id)) {
+        return res.status(400).json({ message: "ID inválido" });
+    }
+    // Verifica se o ID existe
+    const categoriaExistente = await categoria.findByPk(req.params.id);
+    if (!categoriaExistente) {
+        return res.status(404).json({ message: "Categoria não encontrada" });
+    }
+    // Exclui o registro
     await categoria.destroy({
         where: { id: req.params.id }
     })
